Collapse 'gcloud.logging._helpers' into '.entries'.
'.entries' is the only remaining client.
